+++ csvtk/cmd/mutate3.go
@@ -60,12 +60,17 @@ Supported Operators:
   Comparison: > >= < <= == !=
   Logical: not ! and && or ||
   String: + contains startsWith endsWith
+          trim trimPrefix trimSuffix upper lower split splitAfter replace
+          repeat indexOf lastIndexOf hasPrefix hasSuffix
+  Membership: [] . ?. in
   Regex: matches
   Range: ..
   Slice: [:]
   Pipe: |
   Ternary conditional: ? :
   Null coalescence: ??
+  Date: now duration date timezone
+  Number: max min abs ceil floor round
 
 Supported Literals:
 
@@ -81,7 +86,7 @@ See Expr language definition link for documentation on built-in functions.
 
 Custom functions:
   - ulen(), length of unicode strings/width of unicode strings rendered
-    to a terminal, e.g., len("沈伟")==6, ulen("沈伟")==4
+    to a terminal, e.g., len("沈伟")==2, ulen("沈伟")==4
 `,
 	Run: func(cmd *cobra.Command, args []string) {
 		config := getConfigs(cmd)