+++ CHANGELOG.md
@@ -3,8 +3,10 @@
     - `csvtk filter/filter2/mutate2/mutate3/sort/summary/round`:
         - fix wrongly converting `.` to the number zero. [#359](https://github.com/shenwei356/csvtk/issues/359)
     - `csvtk cut`:
-        - suppor `-Z/--show-row-number`. [#360](https://github.com/shenwei356/csvtk/issues/360)
+        - support `-Z/--show-row-number`. [#360](https://github.com/shenwei356/csvtk/issues/360)
         - if there's only one input file, it will be directly outputted.
+      - `csvtk mutate2`:
+        - fix the function `len` for Unicode strings. [#368](https://github.com/shenwei356/csvtk/issues/368)
 - [csvtk v0.37.0](https://github.com/shenwei356/csvtk/releases/tag/v0.37.0)
 [![Github Releases (by Release)](https://img.shields.io/github/downloads/shenwei356/csvtk/v0.37.0/total.svg)](https://github.com/shenwei356/csvtk/releases/tag/v0.37.0)
     - `csvtk`: